fixed as per rowAvg = float( sum(pressures) / (elementsCount - 1) ) in notebook, too

+++ Ex_session_03/code/ex3.2/ex3.2.py
@@ -13,7 +13,7 @@ def averagePressureFor(line):
 	if elementsCount > 0:
 		pressures = pressuresFrom(elements)
 		# print(pressures)
-		rowAvg = sum(pressures) / (elementsCount - 1)
+		rowAvg = float( sum(pressures) / (elementsCount - 1) )
 	else:
 		rowAvg = 0
 	return rowAvg
